backend -> architecture renames
