+++ nobodywho/core/src/stt/architecture.rs
@@ -1,12 +1,12 @@
 use crate::errors::SttError;
 use crate::onnx::Device;
 use crate::stream::StreamOutput;
-use crate::stt::{audio, backends, AudioInput, SttConfig};
+use crate::stt::{architectures, audio, AudioInput, SttConfig};
 use std::time::Instant;
 use tokio::sync::mpsc::UnboundedSender;
 use tracing::info;
 
-pub(super) trait SttBackendImpl: Send {
+pub(super) trait SttArchitectureImpl: Send {
     /// Transcribe a single 30-second window of 16 kHz mono f32 samples.
     /// `on_token` is called with each decoded token piece as it is generated.
     fn transcribe_window(
@@ -16,21 +16,21 @@ pub(super) trait SttBackendImpl: Send {
     ) -> Result<String, SttError>;
 }
 
-pub(super) fn load_backend(
+pub(super) fn load_architecture(
     config: SttConfig,
     device: Device,
-) -> Result<Box<dyn SttBackendImpl>, SttError> {
+) -> Result<Box<dyn SttArchitectureImpl>, SttError> {
     match config {
         SttConfig::Whisper(config) => {
             let init_start = Instant::now();
-            let backend = backends::WhisperBackend::new(
+            let architecture = architectures::WhisperBackend::new(
                 &config.source,
                 config.language.as_deref(),
                 &config.quantization,
                 device,
             )?;
             info!(elapsed = ?init_start.elapsed(), "Initialized Whisper STT");
-            Ok(Box::new(backend))
+            Ok(Box::new(architecture))
         }
     }
 }
@@ -48,7 +48,7 @@ fn decode_input(input: AudioInput) -> Result<Vec<Vec<f32>>, SttError> {
 }
 
 pub(super) fn transcribe_sync(
-    backend: &mut dyn SttBackendImpl,
+    architecture: &mut dyn SttArchitectureImpl,
     input: AudioInput,
 ) -> Result<String, SttError> {
     let start = Instant::now();
@@ -57,7 +57,7 @@ pub(super) fn transcribe_sync(
 
     let mut parts: Vec<String> = Vec::with_capacity(n_windows);
     for (i, window) in windows.into_iter().enumerate() {
-        let text = backend.transcribe_window(&window, &mut |_| {})?;
+        let text = architecture.transcribe_window(&window, &mut |_| {})?;
         info!(window = i + 1, total = n_windows, text = %text, "Transcribed window");
         if !text.trim().is_empty() {
             parts.push(text.trim().to_string());
@@ -70,17 +70,17 @@ pub(super) fn transcribe_sync(
 }
 
 pub(super) fn transcribe_streaming(
-    backend: &mut dyn SttBackendImpl,
+    architecture: &mut dyn SttArchitectureImpl,
     input: AudioInput,
     tx: UnboundedSender<StreamOutput<SttError>>,
 ) {
-    if let Err(e) = do_transcribe_streaming(backend, input, &tx) {
+    if let Err(e) = do_transcribe_streaming(architecture, input, &tx) {
         let _ = tx.send(StreamOutput::Error(e));
     }
 }
 
 fn do_transcribe_streaming(
-    backend: &mut dyn SttBackendImpl,
+    architecture: &mut dyn SttArchitectureImpl,
     input: AudioInput,
     tx: &UnboundedSender<StreamOutput<SttError>>,
 ) -> Result<(), SttError> {
@@ -88,7 +88,7 @@ fn do_transcribe_streaming(
     let mut full_transcript = String::new();
 
     for window in windows {
-        let text = backend.transcribe_window(&window, &mut |piece| {
+        let text = architecture.transcribe_window(&window, &mut |piece| {
             let _ = tx.send(StreamOutput::Token(piece));
         })?;
         if !text.trim().is_empty() {
